# workspaces/backend/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage notebooks in a Kubernetes cluster.\nFor more information, visit https://www.kubeflow.org/docs/components/notebooks/",
        "title": "Kubeflow Notebooks API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/namespaces": {
            "get": {
                "description": "Returns a list of all namespaces in the cluster.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "List namespaces",
                "operationId": "listNamespaces",
                "responses": {
                    "200": {
                        "description": "Successful namespaces response",
                        "schema": {
                            "$ref": "#/definitions/api.NamespaceListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/persistentvolumeclaims/{namespace}": {
            "get": {
                "description": "Returns a list of persistent volume claims in a specific namespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persistentvolumeclaims"
                ],
                "summary": "List persistent volume claims by namespace",
                "operationId": "listPVCs",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "my-namespace",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful PVCs response",
                        "schema": {
                            "$ref": "#/definitions/api.PVCListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new persistent volume claim in the specified namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persistentvolumeclaims"
                ],
                "summary": "Create persistent volume claim",
                "operationId": "createPVC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PVC creation request",
                        "name": "pvc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PVCCreateEnvelope"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PVC created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.PVCCreateEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "PVC already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large. The request body is too large.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type. Content-Type header is not correct.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/persistentvolumeclaims/{namespace}/{name}": {
            "delete": {
                "description": "Deletes a specific persistent volume claim identified by namespace and name.",
                "tags": [
                    "persistentvolumeclaims"
                ],
                "summary": "Deletes a persistent volume claim",
                "operationId": "deletePVC",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "my-namespace",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-example": "my-pvc",
                        "description": "PVC name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "PVC not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ConflictError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "A human-readable description of the cause of the error.\nThis field may be presented as-is to a reader.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin indicates where the conflict error originated.\nIf value is empty, the origin is unknown.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCauseOrigin"
                        }
                    ]
                }
            }
        },
        "api.ErrorCause": {
            "type": "object",
            "properties": {
                "conflict_cause": {
                    "description": "ConflictCauses contains details about conflict errors that caused the request to fail.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ConflictError"
                    }
                },
                "validation_errors": {
                    "description": "ValidationErrors contains details about validation errors that caused the request to fail.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ValidationError"
                    }
                }
            }
        },
...
        "api.ErrorEnvelope": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.HTTPError"
                }
            }
        },
        "api.HTTPError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "cause": {
                    "description": "Cause contains detailed information about the cause of the error.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCause"
                        }
                    ]
                },
                "code": {
                    "description": "Code is a string representation of the HTTP status code.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a human-readable description of the error.",
                    "type": "string"
                }
            }
        },
        "api.NamespaceListEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/namespaces.Namespace"
                    }
                }
            }
        },
        "api.PVCCreateEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/pvcs.PVCCreate"
                }
            }
        },
        "api.PVCListEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvcs.PVCListItem"
                    }
                }
            }
        },
...
        "api.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "The field of the resource that has caused this error, as named by its JSON serialization.\nMay include dot and postfix notation for nested attributes.\nArrays are zero-indexed.\nFields may appear more than once in an array of causes due to fields having multiple errors.\n\nExamples:\n  \"name\" - the field \"name\" on the current resource\n  \"items[0].name\" - the field \"name\" on the first array entry in \"items\"",
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the cause of the error.\nThis field may be presented as-is to a reader.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin indicates where the validation error originated.\nIf value is empty, the origin is unknown.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCauseOrigin"
                        }
                    ]
                },
                "type": {
                    "description": "A machine-readable description of the cause of the error.\nIf value is empty, there is no information available.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/field.ErrorType"
                        }
                    ]
                }
            }
        },
...
        "common.Audit": {
            "type": "object",
            "required": [
                "createdAt",
                "createdBy",
                "deletedAt",
                "updatedAt",
                "updatedBy"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "namespaces.Namespace": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "pvcs.PVCCreate": {
            "type": "object",
            "required": [
                "accessModes",
                "name",
                "requests",
                "storageClassName"
            ],
            "properties": {
                "accessModes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.PersistentVolumeAccessMode"
                    }
                },
                "name": {
                    "type": "string"
                },
                "requests": {
                    "$ref": "#/definitions/pvcs.StorageRequests"
                },
                "storageClassName": {
                    "type": "string"
                }
            }
        },
        "pvcs.PVCListItem": {
            "type": "object",
            "required": [
                "audit",
                "canMount",
                "canUpdate",
                "name",
                "pods",
                "pvcSpec",
                "workspaces"
            ],
            "properties": {
                "audit": {
                    "$ref": "#/definitions/common.Audit"
                },
                "canMount": {
                    "type": "boolean"
                },
                "canUpdate": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "pods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvcs.PodInfo"
                    }
                },
                "pv": {
                    "description": "This field is nil until a PV is bound to the PVC.\nhttps://kubernetes.io/docs/concepts/storage/persistent-volumes/#binding",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvcs.PVInfo"
                        }
                    ]
                },
                "pvcSpec": {
                    "$ref": "#/definitions/pvcs.PVCSpec"
                },
                "workspaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvcs.WorkspaceInfo"
                    }
                }
            }
        },
        "pvcs.PVCSpec": {
            "type": "object",
            "required": [
                "accessModes",
                "requests",
                "storageClassName",
                "volumeMode"
            ],
            "properties": {
                "accessModes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.PersistentVolumeAccessMode"
                    }
                },
                "requests": {
                    "$ref": "#/definitions/pvcs.StorageRequests"
                },
                "storageClassName": {
                    "description": "This field may be an empty string in two cases:\n1. The PVC is requesting the default storage class, and it has not been bound to a PV yet.\n2. The PVC is explicitly requesting a PV with no storage class (i.e. manual or out-of-band binding).",
                    "type": "string"
                },
                "volumeMode": {
                    "$ref": "#/definitions/v1.PersistentVolumeMode"
                }
            }
        },
        "pvcs.PVInfo": {
            "type": "object",
            "required": [
                "accessModes",
                "name",
                "persistentVolumeReclaimPolicy",
                "volumeMode"
            ],
            "properties": {
                "accessModes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.PersistentVolumeAccessMode"
                    }
                },
                "name": {
                    "type": "string"
                },
                "persistentVolumeReclaimPolicy": {
                    "$ref": "#/definitions/v1.PersistentVolumeReclaimPolicy"
                },
                "storageClass": {
                    "description": "This field should only be nil if the bound PV does not have a storage class (i.e. manual or out-of-band binding).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvcs.StorageClassInfo"
                        }
                    ]
                },
                "volumeMode": {
                    "$ref": "#/definitions/v1.PersistentVolumeMode"
                }
            }
        },
        "pvcs.PodInfo": {
            "type": "object",
            "required": [
                "name",
                "phase"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "node": {
                    "$ref": "#/definitions/pvcs.PodNode"
                },
                "phase": {
                    "$ref": "#/definitions/v1.PodPhase"
                }
            }
        },
        "pvcs.PodNode": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pvcs.StorageRequests": {
            "type": "object",
            "required": [
                "storage"
            ],
            "properties": {
                "storage": {
                    "type": "string"
                }
            }
        },
        "pvcs.WorkspaceInfo": {
            "type": "object",
            "required": [
                "name",
                "state",
                "stateMessage"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "podTemplatePod": {
                    "$ref": "#/definitions/pvcs.PodTemplatePod"
                },
                "state": {
                    "$ref": "#/definitions/v1beta1.WorkspaceState"
                },
                "stateMessage": {
                    "type": "string"
                }
            }
        },
...
        "v1.PersistentVolumeAccessMode": {
            "type": "string",
            "enum": [
                "ReadWriteOnce",
                "ReadOnlyMany",
                "ReadWriteMany",
                "ReadWriteOncePod"
            ],
            "x-enum-varnames": [
                "ReadWriteOnce",
                "ReadOnlyMany",
                "ReadWriteMany",
                "ReadWriteOncePod"
            ]
        },
        "v1.PersistentVolumeMode": {
            "type": "string",
            "enum": [
                "Block",
                "Filesystem"
            ],
            "x-enum-varnames": [
                "PersistentVolumeBlock",
                "PersistentVolumeFilesystem"
            ]
        },
        "v1.PersistentVolumeReclaimPolicy": {
            "type": "string",
            "enum": [
                "Recycle",
                "Delete",
                "Retain"
            ],
            "x-enum-varnames": [
                "PersistentVolumeReclaimRecycle",
                "PersistentVolumeReclaimDelete",
                "PersistentVolumeReclaimRetain"
            ]
        },
        "v1.PodPhase": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed",
                "Unknown"
            ],
            "x-enum-varnames": [
                "PodPending",
                "PodRunning",
                "PodSucceeded",
                "PodFailed",
                "PodUnknown"
            ]
        },
...
    }
}